client_service/requirements.txt
# Requirements for the client_service

## Core runtime / Web framework
fastapi==0.117.1
fastapi-mcp==0.4.0
uvicorn==0.36.0
asyncpg==0.30.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.9
pydantic[email]
python-json-logger==4.0.0
requests==2.32.5
beanie==2.0.0
motor==3.7.1
python-jose==3.3.0  
rapidfuzz==3.14.3
